Basic prefix expansion control for Prefix Data
By coming prefixes back to their original country code (China) for
everything but geolocation. The expansion of the prefix still happens
for NANPA countries and one Chinese prefix (+861) to reduce memory usage
for the total number of prefixes (~150,000). This primitive method might
need updating in the future as other prefixes get more data.

When loading the data, it now recursively searches the Prefix Map for
the whole phone number, instead of just the country code. Further
optimisations could be made here in the future to use array keys instead
of array values.